Data initialization change

+++ ProductShop/PetShop.RestAPI/Startup.cs
@@ -1,19 +1,11 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Cors.Infrastructure;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using PetShop.Core.ApplicationService;
@@ -107,7 +99,7 @@ namespace PetShop.RestAPI
 
             app.UseSwaggerUI(c =>
             {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "PetShop API");
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", "ProductShop API");
             });
 
             if (env.IsDevelopment())
@@ -116,11 +108,22 @@ namespace PetShop.RestAPI
                 using (var scope = app.ApplicationServices.CreateScope())
                 {
                     var ctx = scope.ServiceProvider.GetService<ProductShopContext>();
-                    ctx.Database.EnsureDeleted();
-                    ctx.Database.EnsureCreated();
 
-                    InitStaticData dataInitilizer = scope.ServiceProvider.GetRequiredService<InitStaticData>();
-                    dataInitilizer.InitData();
+                    if (Env.IsDevelopment())
+                    {
+                        ctx.Database.EnsureDeleted();
+                        ctx.Database.EnsureCreated();
+
+                        InitStaticData dataInitilizer = scope.ServiceProvider.GetRequiredService<InitStaticData>();
+                        dataInitilizer.InitData();
+                    }
+                    else
+                    {
+                        {
+                            ctx.Database.EnsureCreated();
+                        }
+                    }
+                    
                 }
             }
